requirements.txt
streamlit
pandas
plotly
tables
matplotlib
python-dotenv
plotly
seaborn
matplotlib